sealing pipeline: Cleanup journal plumbing
